Move request policy fetching to request object factory MODPATRON-66
