+++ src/main/java/org/folio/rest/impl/RequestObjectFactory.java
@@ -8,10 +8,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
+import org.folio.integration.http.VertxOkapiHttpClient;
 import org.folio.rest.jaxrs.model.Hold;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 
+import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 
 class RequestObjectFactory {
@@ -57,7 +59,7 @@ class RequestObjectFactory {
       .thenApply(x -> createRequestPolicyIdCriteria(itemFuture, userFuture, requestTypeParams))
       .thenCompose(this::lookupRequestPolicyId)
       .thenCompose(policyIdResponse ->
-          LookupsUtils.getRequestPolicy(policyIdResponse.getString("requestPolicyId"), okapiHeaders))
+          getRequestPolicy(policyIdResponse.getString("requestPolicyId"), okapiHeaders))
       .thenApply(RequestPolicy::from)
       .thenApply(requestPolicy -> getRequestType(requestPolicy, requestTypeParams.getItemStatus()));
   }
@@ -73,7 +75,24 @@ class RequestObjectFactory {
   }
 
   private CompletableFuture<JsonObject> lookupRequestPolicyId(RequestTypeParameters criteria) {
-    return LookupsUtils.getRequestPolicyId(criteria, okapiHeaders);
+    final var client = new VertxOkapiHttpClient(Vertx.currentContext().owner());
+
+    final var queryParameters = Map.of(
+      "item_type_id", criteria.getItemMaterialTypeId(),
+      "loan_type_id", criteria.getItemLoanTypeId(),
+      "patron_type_id", criteria.getPatronGroupId(),
+      "location_id", criteria.getItemLocationId());
+
+    return client.get("/circulation/rules/request-policy", queryParameters, okapiHeaders)
+      .thenApply(LookupsUtils::verifyAndExtractBody);
+  }
+
+  private CompletableFuture<JsonObject> getRequestPolicy(String requestPolicyId, Map<String, String> okapiHeaders) {
+    final var client = new VertxOkapiHttpClient(Vertx.currentContext().owner());
+
+    return client.get("/request-policy-storage/request-policies/" + requestPolicyId,
+        Map.of(), okapiHeaders)
+      .thenApply(LookupsUtils::verifyAndExtractBody);
   }
 
   private RequestTypeParameters createRequestPolicyIdCriteria(CompletableFuture<JsonObject> itemFuture,
